requirements.txt
langchain
langchain-google-vertexai
langchain-anthropic
langgraph
sentence-transformers
chromadb
faiss-cpu
networkx
matplotlib
plotly
dash
pytest
gensim
pytrends
pandas
requests
stripe
paypalrestsdk
beautifulsoup4
httpx
pyyaml
prometheus-client
opentelemetry-api
opentelemetry-sdk
umap-learn
hdbscan
pytest-asyncio
